+++ crates/paradencer-ingress/tests/turbine_integration.rs
@@ -68,7 +68,7 @@ fn test_turbine_tree_construction_small_network() {
 
     let validators: Vec<_> = (1..=10)
         .map(|i| {
-            create_validator_with_addr(create_node_id(i), 1000 * i as u64, 127, 0, 0, 1, 8000 + i)
+            create_validator_with_addr(create_node_id(i), 1000 * i as u64, 127, 0, 0, 1, 8000 + i as u16)
         })
         .collect();
 
@@ -157,7 +157,7 @@ fn test_turbine_tree_stake_weighting_distribution() {
             0,
             0,
             1,
-            8000 + i,
+            8000 + i as u16,
         ));
     }
 
@@ -170,7 +170,7 @@ fn test_turbine_tree_stake_weighting_distribution() {
             0,
             0,
             1,
-            8000 + i,
+            8000 + i as u16,
         ));
     }
 
@@ -217,7 +217,7 @@ fn test_turbine_tree_network_proximity() {
             168,
             1,
             i,
-            8000 + i,
+            8000 + i as u16,
         ));
     }
 
@@ -230,7 +230,7 @@ fn test_turbine_tree_network_proximity() {
             168,
             2,
             i - 10,
-            8000 + i,
+            8000 + i as u16,
         ));
     }
 
@@ -243,7 +243,7 @@ fn test_turbine_tree_network_proximity() {
             0,
             0,
             i - 20,
-            8000 + i,
+            8000 + i as u16,
         ));
     }
 
@@ -395,7 +395,7 @@ async fn test_broadcaster_with_tree() {
 
     let validators: Vec<_> = (1..=20)
         .map(|i| {
-            create_validator_with_addr(create_node_id(i), 1000 * i as u64, 127, 0, 0, 1, 8000 + i)
+            create_validator_with_addr(create_node_id(i), 1000 * i as u64, 127, 0, 0, 1, 8000 + i as u16)
         })
         .collect();
 
@@ -506,7 +506,7 @@ fn test_tree_rebuild_detection() {
     let root_contact = create_contact_info(root_id, 127, 0, 0, 1, 8000);
     let validators: Vec<_> = (1..=10)
         .map(|i| {
-            create_validator_with_addr(create_node_id(i), 1000 * i as u64, 127, 0, 0, 1, 8000 + i)
+            create_validator_with_addr(create_node_id(i), 1000 * i as u64, 127, 0, 0, 1, 8000 + i as u16)
         })
         .collect();
 
@@ -530,7 +530,7 @@ fn test_layer_distribution_balance() {
 
     let validators: Vec<_> = (1..=100)
         .map(|i| {
-            create_validator_with_addr(create_node_id(i), 1000 * i as u64, 127, 0, 0, 1, 8000 + i)
+            create_validator_with_addr(create_node_id(i), 1000 * i as u64, 127, 0, 0, 1, 8000 + i as u16)
         })
         .collect();
 
@@ -603,7 +603,7 @@ fn test_tree_parent_child_consistency() {
 
     let validators: Vec<_> = (1..=50)
         .map(|i| {
-            create_validator_with_addr(create_node_id(i), 1000 * i as u64, 127, 0, 0, 1, 8000 + i)
+            create_validator_with_addr(create_node_id(i), 1000 * i as u64, 127, 0, 0, 1, 8000 + i as u16)
         })
         .collect();
 
@@ -649,7 +649,7 @@ fn test_multiple_tree_rebuilds() {
 
     let validators: Vec<_> = (1..=20)
         .map(|i| {
-            create_validator_with_addr(create_node_id(i), 1000 * i as u64, 127, 0, 0, 1, 8000 + i)
+            create_validator_with_addr(create_node_id(i), 1000 * i as u64, 127, 0, 0, 1, 8000 + i as u16)
         })
         .collect();
